NF: RescheduleRepositionReset together
