+++ AnkiDroid/src/main/java/com/ichi2/anki/Reviewer.java
@@ -465,7 +465,7 @@ public class Reviewer extends AbstractFlashcardViewer {
     private void showRescheduleCardDialog() {
         Consumer<Integer> runnable = days ->
             CollectionTask.launchCollectionTask(DISMISS_MULTI, mRescheduleCardHandler,
-                    new TaskData(new CollectionTask.RescheduleRepositionReset(new long[]{mCurrentCard.getId()},
+                    new TaskData(new CardBrowser.RescheduleRepositionReset(new long[]{mCurrentCard.getId()},
                     Collection.DismissType.RESCHEDULE_CARDS, days))
             );
         RescheduleDialog dialog = RescheduleDialog.rescheduleSingleCard(getResources(), mCurrentCard, runnable);
@@ -485,7 +485,7 @@ public class Reviewer extends AbstractFlashcardViewer {
         Runnable confirm = () -> {
             Timber.i("NoteEditor:: ResetProgress button pressed");
             CollectionTask.launchCollectionTask(DISMISS_MULTI, mResetProgressCardHandler,
-                    new TaskData(new CollectionTask.RescheduleRepositionReset(new long[]{mCurrentCard.getId()}, Collection.DismissType.RESET_CARDS)));
+                    new TaskData(new CardBrowser.RescheduleRepositionReset(new long[]{mCurrentCard.getId()}, Collection.DismissType.RESET_CARDS)));
         };
         dialog.setConfirm(confirm);
         showDialogFragment(dialog);
