+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionTask.java
@@ -460,27 +460,6 @@ public class CollectionTask extends BaseAsyncTask<TaskData, TaskData, TaskData> 
         }
     }
 
-
-    private static class UndoRepositionRescheduleResetCards extends Undoable {
-        private final Card[] cards_copied;
-
-
-        public UndoRepositionRescheduleResetCards(Collection.DismissType type, Card[] cards_copied) {
-            super(type);
-            this.cards_copied = cards_copied;
-        }
-
-
-        public long undo(Collection col) {
-            Timber.i("Undoing action of type %s on %d cards", getDismissType(), cards_copied.length);
-            for (int i = 0; i < cards_copied.length; i++) {
-                Card card = cards_copied[i];
-                card.flush(false);
-            }
-            return NO_REVIEW;
-        }
-    }
-
     public static class ChangeDeckMulti extends DismissMulti {
         private long mNewDid;
         public ChangeDeckMulti(long[] cardIds, long newDid) {
@@ -537,69 +516,6 @@ public class CollectionTask extends BaseAsyncTask<TaskData, TaskData, TaskData> 
             // mark undo for all at once
             col.markUndo(new UndoChangeDeckMulti(cards, originalDids));
             return null;
-        }
-    }
-
-    public static class RescheduleRepositionReset extends DismissMulti {
-        private final Collection.DismissType mType;
-        private final int mData;
-
-        public RescheduleRepositionReset(long[] cardDids, Collection.DismissType type) {
-            this(cardDids, type, 0);
-        }
-
-        public RescheduleRepositionReset(long[] cardDids, Collection.DismissType type, int data) {
-            super(cardDids);
-            mType = type;
-            mData = data;
-        }
-
-        public TaskData actualBackground(CollectionTask task, Card[] cards) {
-            Collection col = task.getCol();
-            AbstractSched sched = col.getSched();
-            switch(mType) {
-            case RESCHEDULE_CARDS:
-            case REPOSITION_CARDS:
-            case RESET_CARDS: 
-                // collect undo information, sensitive to memory pressure, same for all 3 cases
-                try {
-                    Timber.d("Saving undo information of type %s on %d cards", mType, cards.length);
-                    col.markUndo(new UndoRepositionRescheduleResetCards(mType, deepCopyCardArray(task, cards)));
-                } catch (CancellationException ce) {
-                    Timber.i(ce, "Cancelled while handling type %s, skipping undo", mType);
-                }
-                switch (mType) {
-                case RESCHEDULE_CARDS:
-                    sched.reschedCards(getCardIds(), mData, mData);
-                    break;
-                case REPOSITION_CARDS:
-                    sched.sortCards(getCardIds(), mData, 1, false, true);
-                    break;
-                case RESET_CARDS:
-                    sched.forgetCards(getCardIds());
-                    break;
-                }
-                // In all cases schedule a new card so Reviewer doesn't sit on the old one
-                col.reset();
-                task.doProgress(new TaskData(sched.getCard(), 0));
-            }
-            return null;
-        }
-
-        private Card[] deepCopyCardArray(CollectionTask task, Card[] originals) throws CancellationException {
-            Collection col = CollectionHelper.getInstance().getCol(AnkiDroidApp.getInstance());
-            Card[] copies = new Card[originals.length];
-            for (int i = 0; i < originals.length; i++) {
-                if (task.isCancelled()) {
-                    Timber.i("Cancelled during deep copy, probably memory pressure?");
-                    throw new CancellationException("Cancelled during deep copy");
-                }
-
-                // TODO: the performance-naive implementation loads from database instead of working in memory
-                // the high performance version would implement .clone() on Card and test it well
-                copies[i] = new Card(col, originals[i].getId());
-            }
-            return copies;
         }
     }
